+++ applications/frontend/config/email.php
@@ -7,12 +7,24 @@
 | Configuration of outgoing mail server.
 | */
 
+// Mandrill
+$config['protocol'] = 'smtp';
+$config['smtp_host'] = 'smtp.mandrillapp.com';
+$config['smtp_port'] = '587';
+$config['smtp_timeout'] = '30';
+$config['smtp_user'] = 'username';
+$config['smtp_pass'] = 'password';
+
+// Mailgun
+/*
 $config['protocol'] = 'smtp';
 $config['smtp_host'] = 'ssl://smtp.mailgun.org';
 $config['smtp_port'] = '465';
 $config['smtp_timeout'] = '30';
 $config['smtp_user'] = 'postmaster@example.com';
 $config['smtp_pass'] = 'password';
+*/
+
 $config['charset'] = 'utf-8';
 $config['mailtype'] = 'html';
 $config['wordwrap'] = TRUE;