+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 #### v 0.7.0 (Date: 2015-03-18)
 - **Frontend**
 	- Added multilingual support
+	- Added config sample for email via Mandrill
 
 #### v 0.6.1 (Date: 2015-03-10)
 - **Backend**