+++ CHANGELOG.md
@@ -1,4 +1,10 @@
 
+#### v 0.7.1 (Date: 2015-04-17)
+- **General**
+	- Updated CodeIgniter to v2.2.2
+	- Updated AdminLTE to v2.0.5
+	- Updated Bootstrap to v3.3.4
+
 #### v 0.7.0 (Date: 2015-03-18)
 - **Frontend**
 	- Added multilingual support